requirements.txt
openai-whisper==20231117
torch==2.1.0
numpy==1.24.3
tqdm==4.66.1
requests==2.31.0
librosa==0.10.1 